% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_axis.R
\name{rm_axis}
\alias{rm_axis}
\title{Remove an axis from a ggplot}
\usage{
rm_axis(axis, line_too = TRUE)
}
\arguments{
\item{axis}{The axis to remove ("x" or "y")}

\item{line_too}{Whether to remove the axis line too (i.e. not just the title,
ticks and text)}
}
\description{
Remove an axis from a ggplot
}
\examples{

data <- tibble::tibble(x = 1:3, z = 3:1)

ggplot2::ggplot(data, ggplot2::aes(x = x, y = x)) +
  ggplot2::geom_line() +
  rm_axis("x")

}
