% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_strips.R
\name{rm_strips}
\alias{rm_strips}
\title{Remove facet strips from a ggplot}
\usage{
rm_strips(axis = "xy")
}
\arguments{
\item{axis}{Which axis to remove facets along? Either of "x", "y" or "xy".}
}
\description{
Remove facet strips from a ggplot
}
\examples{

data <- tibble::tibble(x = 1:3, y = 1, z = 3:1)

ggplot2::ggplot(data, ggplot2::aes(x = x, y = x)) +
  ggplot2::geom_line() +
  ggplot2::facet_grid(. ~ y) +
  rm_strips()

}
